Massive Bug Fix! Read Descrption`
Massive Bug Fix!
this fixed many bugs, like the xyz cords being slow, and added water and removed the 1 block deep that was always in the sky, made biomes better and made the dungeons a lot less common!
We are Planing to add many new things! like fixing the swimming, and make it like the old MC, or adding Many more blocks (altho its gonna need a search bar, which is not gonna be easy!) Like Flowers!